requirements.txt
segment-anything
opencv-python
numpy
pillow
simple-lama-inpainting
